refactor(color): * split up color into specialized modules * renamed a few color functions for coherence * added/ updated docstrings where necessary * updated tests
